Update README.md
This change adds a sentence to introduce the content of the repository. The key purpose is to give a name to the items that live in this repository so we can common vocabulary. 

Understanding "guides" isn't ideal since they're meant to be very short. They more like "quick notes" on related technologies. For example "we should create write a quick note to the Understanding repository".